Restructured menus, moved SAVE and EXPORT to top-level (at least for now) and clarified instructions
